Add more routes

+++ EMS/routes/web.php
@@ -7,8 +7,6 @@ use App\Http\Controllers\EventController;
 use App\Http\Controllers\ExportController;
 use App\Http\Controllers\RegistrationController;
 use App\Http\Controllers\UserController;
-use App\Models\Categories;
-use App\Models\Events;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -27,46 +25,46 @@ Route::post("/logout",[AdminController::class,'logout'])->name('logout');
 
 //Category Routes
 // Route::view("/categories","categories");
-Route::get("/categories",[CategoryController::class,'listAllCatgory'])->name('category.list');
+Route::get("/categories",[CategoryController::class,'listAllCatgory'])->name('category.list')->middleware('admin');
 
-Route::get("/addcategories",[CategoryController::class,'addCategoryView'])->name('category.view');
-Route::post("/categories",[CategoryController::class,'addCategory'])->name('category.add');
+Route::get("/addcategories",[CategoryController::class,'addCategoryView'])->name('category.view')->middleware('admin');
+Route::post("/categories",[CategoryController::class,'addCategory'])->name('category.add')->middleware('admin');
 
 //Update Category
-Route::get("/updatecategories/{id}",[CategoryController::class,'updateCategoryView'])->name('category.uView');
-Route::put("/updatecategories/{id}",[CategoryController::class,'updateCategory'])->name('category.edit');
+Route::get("/updatecategories/{id}",[CategoryController::class,'updateCategoryView'])->name('category.uView')->middleware('admin');
+Route::put("/updatecategories/{id}",[CategoryController::class,'updateCategory'])->name('category.edit')->middleware('admin');
 
 //Delete Category
-Route::delete("/deletecategories/{id}",[CategoryController::class,'deleteCategory'])->name('category.delete');
+Route::delete("/deletecategories/{id}",[CategoryController::class,'deleteCategory'])->name('category.delete')->middleware('admin');
 
 //Event routes
 // Route::view("/events","events")->name('events');
-Route::get("/event",[EventController::class,'index'])->name('events.event');
-Route::get("/events",[EventController::class,'showAllEvent'])->name("event.AllEvents");
-Route::get("/recentevents",[EventController::class,'recentEvent'])->name("event.recentEvents");
+Route::get("/event",[EventController::class,'index'])->name('events.event')->middleware('admin');
+Route::get("/events",[EventController::class,'showAllEvent'])->name("event.AllEvents")->middleware('admin');
+Route::get("/recentevents",[EventController::class,'recentEvent'])->name("event.recentEvents")->middleware('admin');
 
 //Add Events
-Route::view("/upload","uploadevent");
-Route::post("/upload",[EventController::class,'addEvent'])->name('event.uploadEvent');
+Route::view("/upload","uploadevent")->middleware('admin');
+Route::post("/upload",[EventController::class,'addEvent'])->name('event.uploadEvent')->middleware('admin');
 
 //Update Events
-Route::get('/updateEvent/{id}',[EventController::class,'updateEventView'])->name('event.updateEventView');
-Route::put("/updateEvent/{id}",[EventController::class,'updateEvent'])->name('event.updateEvent');
+Route::get('/updateEvent/{id}',[EventController::class,'updateEventView'])->name('event.updateEventView')->middleware('admin');
+Route::put("/updateEvent/{id}",[EventController::class,'updateEvent'])->name('event.updateEvent')->middleware('admin');
 //Delete Events
-Route::delete('/deleteevent/{id}',[EventController::class,'deleteEvent'])->name('event.deleteEvent');
+Route::delete('/deleteevent/{id}',[EventController::class,'deleteEvent'])->name('event.deleteEvent')->middleware('admin');
 
 //Search Event
-Route::get('/search',[EventController::class,'searchEvent'])->name('event.searchEvent');
+Route::get('/search',[EventController::class,'searchEvent'])->name('event.searchEvent')->middleware('admin');
 
 //User
 // Route::view("/users","users")->name('user.users');
-Route::get('/users',[UserController::class,'index'])->name('user.users');
-Route::get('/reguser',[UserController::class,'registerdUser'])->name('user.rgusers');
+Route::get('/users',[UserController::class,'index'])->name('user.users')->middleware('admin');
+Route::get('/reguser',[UserController::class,'registerdUser'])->name('user.rgusers')->middleware('admin');
 
 //Registration
-Route::get("/registrations",[RegistrationController::class,'index'])->name('registration');
-Route::post("/registrations",[RegistrationController::class,'filterReg'])->name('registration.filter');
-Route::get("/recentregistration",[RegistrationController::class,'recentReg'])->name("registration.recent");
+Route::get("/registrations",[RegistrationController::class,'index'])->name('registration')->middleware('admin');
+Route::post("/registrations",[RegistrationController::class,'filterReg'])->name('registration.filter')->middleware('admin');
+Route::get("/recentregistration",[RegistrationController::class,'recentReg'])->name("registration.recent")->middleware('admin');
 
 
 //livewire
@@ -75,3 +73,15 @@ Route::view('/dish', 'demo')->name('dish-table');
 //Export CSV
 Route::get('/export',[ExportController::class,'exportCSV'])->name('export.csv');
 Route::get('/exports',[ExportController::class,'exportExcel'])->name('export.excel');
+
+Route::view('/abort','403')->name('abort');
+
+Route::fallback(function(){
+    return view('404');
+});
+
+
+/*
+Route::middleware(['admin'])->group(function () {
+    Route::get('/admin/categories', [CategoryController::class,'listAllCatgory'])->name('admin.categories');
+});*/
